duplex-scan-for-any-printer/requirements.txt
# All Python dependencies
numpy
opencv-python-headless
Pillow
pyvips
PyYAML
watchdog
pyftpdlib
PyMuPDF
reportlab
onnxruntime
withoutbg
deskew

# Telegram bot
python-telegram-bot[http2]>=20.0

# Web UI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
httpx>=0.24.0
